refactoring unit tests code
