+++ ParserTests/samples/VariableExpressionTests.cs
@@ -5,7 +5,7 @@ using sly.parser;
 using sly.parser.generator;
 using Xunit;
 
-namespace ParserTests
+namespace ParserTests.samples
 {
     public class VariableExpressionTests
     {
