+++ ParserTests/samples/WhileTests.cs
@@ -7,7 +7,7 @@ using sly.parser;
 using sly.parser.generator;
 using Xunit;
 
-namespace ParserTests
+namespace ParserTests.samples
 {
     public class WhileTests
     {
